test(onboard): split web search helpers
